fix: move internal header to src/
