+++ cpp/daal/src/algorithms/kmeans/kmeans_lloyd_batch_impl.i
@@ -32,7 +32,7 @@
 #include "src/algorithms/kmeans/kmeans_lloyd_impl.i"
 #include "src/algorithms/kmeans/kmeans_lloyd_postprocessing.h"
 
-#include "services/internal/service_profiler.h"
+#include "src/services/service_profiler.h"
 
 using namespace daal::internal;
 using namespace daal::services::internal;
