+++ cpp/daal/src/algorithms/low_order_moments/low_order_moments_csr_fast_distr_step2_fpt_cpu.cpp
@@ -20,7 +20,7 @@
 //  Implementation of low order moments kernel.
 //--
 */
-#include "services/internal/service_profiler.h"
+#include "src/services/service_profiler.h"
 
 #include "src/algorithms/low_order_moments/low_order_moments_container.h"
 #include "src/algorithms/low_order_moments/low_order_moments_distributed_impl.i"
