+++ cpp/daal/src/algorithms/low_order_moments/low_order_moments_csr_singlepass_online_fpt_cpu.cpp
@@ -21,7 +21,7 @@
 //--
 */
 
-#include "services/internal/service_profiler.h"
+#include "src/services/service_profiler.h"
 
 #include "src/algorithms/low_order_moments/low_order_moments_container.h"
 #include "src/algorithms/low_order_moments/low_order_moments_online_impl.i"
