+++ cpp/daal/src/services/library_version_info.cpp
@@ -26,7 +26,7 @@
 #include "src/services/service_defines.h"
 #include "services/env_detect.h"
 
-#include "services/internal/service_profiler.h"
+#include "src/services/service_profiler.h"
 
 // Initialize for DAAL INFO output in KERNEL_PROFILER
 static bool print_profiler_header = (daal::internal::print_header(), false);
